Move src/utils/media-query-props to src/core/dom
